+++ tests/app/Services/IntentionCode/IntentionCodeServiceTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Services;
+namespace App\Services\IntentionCode;
 
 use App\BaseFunctionalTestCase;
 use App\Models\IntentionCode\IntentionCode;
